Update related charges (from SFDA)
Per SFDA, 186.22(A) is not a charge eligible for relief under Prop 64,
so we are removing it from the list of related charges we look for.

Also, we clarified that the 664 modifier makes all Prop 64 charges
(already implemented) and related charges eligible for relief, so added
that to our matcher.